+++ app.py
@@ -1,6 +1,5 @@
 import flask
-from flask_login import LoginManager, login_user
-from werkzeug.security import generate_password_hash, check_password_hash
+from flask_login import LoginManager
 from dotenv import load_dotenv, find_dotenv
 import os
 from models import Users, db
@@ -27,48 +26,8 @@ login_manager.init_app(app)
 
 
 @login_manager.user_loader
-def load_user(user_name):
-    return Users.query.get(user_name)
-
-
-@app.route("/signup")
-def signup():
-    return flask.render_template("signup.html")
-
-
-@app.route("/signup", methods=["POST"])
-def signup_post():
-    username = flask.request.form.get("username")
-    password = flask.request.form.get("password")
-    user = Users.query.filter_by(username=username).first()
-    if user:
-        pass
-    else:
-        user = Users(username=username,password=generate_password_hash(password, method='sha256'))
-        db.session.add(user)
-        db.session.commit()
-
-    return flask.redirect(flask.url_for("login"))
-
-
-@app.route("/login")
-def login():
-    return flask.render_template("login.html")
-
-
-@app.route("/login", methods=["POST"])
-def login_post():
-    username = flask.request.form.get("username")
-    password = flask.request.form.get("password")
-    user = Users.query.filter_by(username=username).first()
-    if user and check_password_hash(user.password, password):
-        login_user(user)
-        return flask.redirect(flask.url_for("hearth_routes.index"))
-
-    else:
-        flask.flash("You don't have an account please sign up!")
-    return flask.redirect("login")
-
+def load_user(id):
+    return Users.query.get(int(id))
 
 
 app.register_blueprint(account_routes)
