+++ public/login.html
@@ -15,6 +15,6 @@
             {% endif %}
         {% endwith %}
         <br>
-        <a href={{ url_for('signup') }}>Sign Up</a>
+        <a href="{{url}}">Sign Up</a>
     </body>
 </html>
