+++ public/signup.html
@@ -7,6 +7,6 @@
             <input type="password" name="password" placeholder="Your Password" autofocus="" maxlength="30">
             <input type="submit" value="Submit">
         </form>
-        <a href={{ url_for('login') }}>Log In</a>
+        <a href="{{url}}">Log In</a>
     </body>
 </html>
